Rename and move exiting usecase to vets package
