+++ core/src/test/java/net/kemitix/naolo/core/vets/ListAllVetsTest.java
@@ -1,11 +1,13 @@
-package net.kemitix.naolo.core;
+package net.kemitix.naolo.core.vets;
 
 import net.jqwik.api.*;
 import net.jqwik.api.arbitraries.LongArbitrary;
 import net.jqwik.api.arbitraries.SizableArbitrary;
 import net.jqwik.api.arbitraries.StringArbitrary;
+import net.kemitix.naolo.core.StreamZipper;
 import net.kemitix.naolo.entities.VetSpecialisation;
 import net.kemitix.naolo.entities.Veterinarian;
+import net.kemitix.naolo.storage.spi.VetsRepository;
 import org.assertj.core.api.WithAssertions;
 
 import java.util.List;
@@ -13,24 +15,35 @@ import java.util.Set;
 import java.util.concurrent.ExecutionException;
 import java.util.stream.Stream;
 
-import static net.kemitix.naolo.core.VeterinariansListAll.request;
+import static net.kemitix.naolo.core.vets.ListAllVets.request;
 import static org.mockito.BDDMockito.given;
 import static org.mockito.Mockito.mock;
 
-class VeterinariansListAllTest implements WithAssertions {
+public class ListAllVetsTest implements WithAssertions {
 
-    private final VeterinarianRepository repository = mock(VeterinarianRepository.class);
+    private final VetsRepository repository = mock(VetsRepository.class);
 
-    private final VeterinariansListAll useCase = VeterinariansListAll.create(repository);
+    private final ListAllVets useCase = ListAllVets.create(repository);
+
+    @Provide
+    public static Arbitrary<List<Veterinarian>> vets() {
+        final LongArbitrary ids = Arbitraries.longs();
+        final StringArbitrary names = Arbitraries.strings();
+        final SizableArbitrary<Set<VetSpecialisation>> specialisations = Arbitraries.of(VetSpecialisation.class)
+                .set().ofMinSize(0).ofMaxSize(VetSpecialisation.values().length);
+        return Combinators.combine(ids, names, specialisations)
+                .as(Veterinarian::create)
+                .list();
+    }
 
     @Property
-    void listAll(
+    public void listAll(
             @ForAll("vets") final List<Veterinarian> vets
     ) throws ExecutionException, InterruptedException {
         //given
         given(repository.findAll()).willReturn(vets.stream());
         //when
-        final VeterinariansListAll.Response response = useCase.invoke(request()).get();
+        final ListAllVets.Response response = useCase.invoke(request()).get();
         //then
         final Stream<Tuple.Tuple2<Veterinarian, Veterinarian>> zipped =
                 StreamZipper.zip(vets, response.getAllVeterinarians(), Tuple::of);
@@ -42,16 +55,5 @@ class VeterinariansListAllTest implements WithAssertions {
                     assertThat(r.getName()).isEqualTo(s.getName());
                     assertThat(r.getSpecialisations()).isEqualTo(s.getSpecialisations());
                 });
-    }
-
-    @Provide
-    static Arbitrary<List<Veterinarian>> vets() {
-        final LongArbitrary ids = Arbitraries.longs();
-        final StringArbitrary names = Arbitraries.strings();
-        final SizableArbitrary<Set<VetSpecialisation>> specialisations = Arbitraries.of(VetSpecialisation.class)
-                .set().ofMinSize(0).ofMaxSize(VetSpecialisation.values().length);
-        return Combinators.combine(ids, names, specialisations)
-                .as(Veterinarian::create)
-                .list();
     }
 }
